requirements.txt
tensorflow>=2.4.0
keras>=2.4.0
mtcnn>=0.1.0
numpy>=1.19.0
opencv-python>=4.5.0
pandas>=1.2.0
Pillow>=8.0.0
scikit-learn>=0.24.0
pyttsx3>=2.90
torch>=1.8.0
torchvision>=0.9.0
facenet-pytorch